Do some work

+++ priv/repo/migrations/20171118174106_create_ingredients.exs
@@ -4,7 +4,7 @@ defmodule Food.Repo.Migrations.CreateIngredients do
   def change do
     create table(:ingredients) do
       add :name, :string
-      add :amount, :integer
+      add :amount, :float
       add :unit, :string
       add :recipe_id, references("recipes", on_delete: :delete_all)
